w3c-reports/12092014/mac/runned-tests/XMLHttpRequest/send-authentication-cors-basic-setrequestheader-expected.txt

Blocked access to external URL http://www1.localhost:8800/XMLHttpRequest/resources/auth2/corsenabled.py

FAIL XMLHttpRequest: send() - "Basic" authenticated CORS request using setRequestHeader() (expects to succeed) assert_equals: expected "0a795822-ef1d-42e9-b4d0-0cfb93052e35\npass" but got ""(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/XMLHttpRequest/send-authentication-cors-basic-setrequestheader.htm:28:26
step@http://localhost:8800/resources/testharness.js:798:30
onreadystatechange@http://localhost:8800/XMLHttpRequest/send-authentication-cors-basic-setrequestheader.htm:27:20
send@[native code]
http://localhost:8800/XMLHttpRequest/send-authentication-cors-basic-setrequestheader.htm:34:20
step@http://localhost:8800/resources/testharness.js:798:30
global code@http://localhost:8800/XMLHttpRequest/send-authentication-cors-basic-setrequestheader.htm:16:16)
